refactor: MyInfoView(#25)
